backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch==2.1.1
transformers==4.36.0
numpy==1.24.4
python-multipart==0.0.6
python-dotenv==1.0.0
setuptools>=69
wheel>=0.43
google-generativeai>=0.6.0
